ya marca horas fallo en las funciones

+++ Stats/Core/Session.cs
@@ -5,7 +5,7 @@ namespace Stats
 	public class Session
 	{
 		public long start{ get; set;}
-		public int duration{ get; set;}
+		public long duration{ get; set;}
 		public int distance{ get; set;}
 		public int weight{ get; set;}
 	}
